chore: extract controlservice helpers without splitting methods
