+++ internal/controlservice/runtime_config_helpers.go
@@ -1,0 +1,51 @@
+package controlservice
+
+import (
+	"github.com/buildkite/cleanroom/internal/backend"
+	"github.com/buildkite/cleanroom/internal/runtimeconfig"
+)
+
+func resolveBackendName(requested, configuredDefault string) string {
+	if requested != "" {
+		return requested
+	}
+	if configuredDefault != "" {
+		return configuredDefault
+	}
+	return runtimeconfig.DefaultBackendForHost()
+}
+
+func mergeBackendConfig(backendName string, opts executionOptions, cfg runtimeconfig.Config) backend.FirecrackerConfig {
+	out := backend.FirecrackerConfig{
+		BinaryPath:           cfg.Backends.Firecracker.BinaryPath,
+		KernelImagePath:      cfg.Backends.Firecracker.KernelImage,
+		RootFSPath:           cfg.Backends.Firecracker.RootFS,
+		DockerStartupSeconds: cfg.Backends.Firecracker.Services.Docker.StartupTimeoutSeconds,
+		DockerStorageDriver:  cfg.Backends.Firecracker.Services.Docker.StorageDriver,
+		DockerIPTables:       cfg.Backends.Firecracker.Services.Docker.IPTables,
+		PrivilegedMode:       cfg.Backends.Firecracker.PrivilegedMode,
+		PrivilegedHelperPath: cfg.Backends.Firecracker.PrivilegedHelperPath,
+		VCPUs:                cfg.Backends.Firecracker.VCPUs,
+		MemoryMiB:            cfg.Backends.Firecracker.MemoryMiB,
+		GuestCID:             cfg.Backends.Firecracker.GuestCID,
+		GuestPort:            cfg.Backends.Firecracker.GuestPort,
+		LaunchSeconds:        cfg.Backends.Firecracker.LaunchSeconds,
+	}
+	if backendName == "darwin-vz" {
+		out.KernelImagePath = cfg.Backends.DarwinVZ.KernelImage
+		out.RootFSPath = cfg.Backends.DarwinVZ.RootFS
+		out.DockerStartupSeconds = cfg.Backends.DarwinVZ.Services.Docker.StartupTimeoutSeconds
+		out.DockerStorageDriver = cfg.Backends.DarwinVZ.Services.Docker.StorageDriver
+		out.DockerIPTables = cfg.Backends.DarwinVZ.Services.Docker.IPTables
+		out.VCPUs = cfg.Backends.DarwinVZ.VCPUs
+		out.MemoryMiB = cfg.Backends.DarwinVZ.MemoryMiB
+		out.GuestPort = cfg.Backends.DarwinVZ.GuestPort
+		out.LaunchSeconds = cfg.Backends.DarwinVZ.LaunchSeconds
+	}
+
+	out.Launch = true
+	if opts.LaunchSeconds != 0 {
+		out.LaunchSeconds = opts.LaunchSeconds
+	}
+	return out
+}
